+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Show correct client in HostStatusPanel after full reload
 - Correctly refresh client configuration page after full reload
 - Nullpointer Exception when reloading host configs of a depot
+- Also reset netboot product states when the user wants to reset Localboot AND Netboot product states
 
 ## [4.4.0.6] - 2024-09-05
 ### Fixed